Add ansi_colour.h and migrate console colours
Introduce common/include/ansi_colour.h and migrate display and status color handling to use the new ansi::Colours::Colour API. Replace many console color usages across display headers/implementations, jamstapl OLED handler, DMX monitor, flashcode installer, show system time and debug exception; adapt TextStatus to print ANSI sequences. Replace lib-clib perror internals to use stdio (printf/puts). Also include minor style/brace formatting changes and copyright year updates, plus small .cproject adjustments.

+++ lib-osc/include/oscclient.h
@@ -225,7 +225,7 @@ class OscClient {
 
     void static StaticCallbackFunction(const uint8_t* buffer, uint32_t size, uint32_t from_ip, uint16_t from_port) { s_this->Input(buffer, size, from_ip, from_port); }
 
-    bool HandleLedMessage(uint16_t bytes_received);
+    bool HandleLedMessage(uint32_t bytes_received);
 
    private:
     uint16_t port_outgoing_;
